FIR coefficients are now in a constexpr struct containing a float[][]
needed to bring in the source code for arm_sin_f32 and sinTable_f32
because they aren't constexpr's.
needed to initialize the local variables to make compiler happy.
To make the initialization of the float[][] work I found a technique
from stack overflow, to wrap it in a struct,
and initialize in the constructor.

To apply the changes, arm_fir_init_f32() was moved into the constructor,
as it only needs to be initailized once.
Then, when the filter coeffs need to change (ie. cutoff freq changes)
the pCoeff's member of the FIR struct is set to the appropriate
set of coefficients in the constexpr table.