SerialLogger and Monitor Task are now different things
SerialLogger waits for messages on the queue and sends them.
Monitor is the lowest priority and yields to all other tasks.
Audio Task is the highest priority and when it unblocks,
it resets the Monitor (Idle monitor).
Then whenever the Monitor task runs, it counts how much time it gets.