FFT works
add output parameter for spectrum in FFTProcessor.
FFTBlock holds the input data and the spectrum,
and will calculate spectrum max.

Need to zero pad the input data every time because the
ARM FFT processor will overwrite it.

It does work but the frequency bins are off.
Formula for frequency bin is sampling frequency divided by fft size.
So to get 1Hz bins, need to downsample and/or increase fft size,
but fft size is already at the max for ARM DSP, 4096.
So should downsample to roughly 4096, ie. 4000 = 48000/12
So downsample by 12 with an anti alias filter at 4000Hz.