Large refactorings
Moved more stuff out of header files and into cpp files,
because I realized it's harder to test pieces when you have
too many dependencies in header files.

Tried to get CircularFIRProcessor to beat ARM FIR but I could not
come close.
But atleast it became testable and the FIRCoefficients class
came out of it.  And DelayBuffer is in a new file.
Still havent figured out how to drive it from the
app.  And it probably slows things down even more.