make FFTBlock buildable by itself by defining my_abs()
interestingly removing the ARMDSPFIRProcessor from AudioProcessor.cpp caused
the FFTBlock header to fail, due to the inclusion of arm_math going away,
and FFTBlock was using the abs() from it.
creating a FFTBlock.cpp containing only #include FFTBlock.hpp revealed indeed
the header doesn't build by itself.

to make it buildable on the host, could not use arm_math.h
so the easiest thing in this case was just to define abs() because it's trivial.

funny, I made a mistake putting a > instead of a < and it caused a unit test to break.
THANK YOU , UNIT TESTS.