DelayBlock interpolates delayNumSamples over the buffer length.  So there isn't a discontinuity.
The delayNumSamples in between buffers being processed can change
by more than 1, which can result in severe discontiunities ie. clicks and pops.
so, a simple way to do it is to ramp the delayNumSamples from the last
value to the new value, over the length of the buffer,
ending at the new value.